refactored - works except for legend
