+++ src/components/legend/legend.js
@@ -1,11 +1,17 @@
-import React, { Fragment } from 'react';
+import React, { useState, Fragment, useEffect } from 'react';
 import { Avatar, Box, Card, Stack } from '@mui/joy';
 import { useLayers } from '@context';
 import { getNamespacedEnvParam } from "@utils";
+import { useLocalStorage } from '@hooks';
+import SldStyleParser from 'geostyler-sld-parser';
 
 import Draggable from "react-draggable";
 import { Resizable } from "react-resizable";
 import "react-resizable/css/styles.css";
+
+const MAXELE = 'maxele';
+const MAXWVEL = 'maxwvel';
+const SWAN = 'swan';
 
 export const MapLegend = () => {
     // set correct map styles for layer name
@@ -22,6 +28,13 @@ export const MapLegend = () => {
         'maxele_level_downscaled_epsg4326': 'maxele_v3_short_legend_style',
         'hec_ras_water_surface': 'maxele_style_v3ui'
     };
+
+    const [storedMaxeleStyle, setStoredMaxeleStyle] = useLocalStorage(MAXELE);
+    const [storedMaxwvelStyle, setStoredMaxwvelStyle] = useLocalStorage(MAXWVEL);
+    const [storedSwanStyle, setStoredSwanStyle] = useLocalStorage(SWAN);
+    const [currentStyle, setCurrentStyle] = useState("");
+
+    const sldParser = new SldStyleParser();
 
     const {
         defaultModelLayers,
@@ -40,13 +53,40 @@ export const MapLegend = () => {
         // now build appropriate url for retrieving the legend graphic
         const workspace = legendLayer.layers.split(':')[0];
         const layerName = legendLayer.layers.split(':')[1];
-        const style = layerStyles[layerName.substring(layerName.indexOf('_')+1)];
+        //const style = layerStyles[layerName.substring(layerName.indexOf('_')+1)];
+        let style = "";
+        if (legendLayer.properties.product_type.includes("maxwvel")) {
+            style = storedMaxwvelStyle;
+        }
+        else 
+        if (legendLayer.properties.product_type.includes("swan")) {
+            style = storedSwanStyle;
+        }
+        else { // maxele 
+            style = storedMaxeleStyle;
+        }
 
+        // add the layer name to the style
+        sldParser
+        .readStyle(style)
+        .then((geostylerStyle) => {
+            geostylerStyle.output.name = legendLayer.layers;
+            const colormapEntries = [...geostylerStyle.output.rules[0].symbolizers[0].colorMap.colorMapEntries];
+            const newcolormapEntries = colormapEntries.reverse();
+            geostylerStyle.output.rules[0].symbolizers[0].colorMap.colorMapEntries = newcolormapEntries;
+            sldParser.writeStyle(geostylerStyle.output)
+            .then((sldStyle) => {
+                setCurrentStyle(sldStyle.output);
+            });
+        }); 
+
+        console.log(style);
         legendUrl = `${ getNamespacedEnvParam('REACT_APP_GS_DATA_URL') }` +
                     workspace + "/" +
                     "ows?service=WMS&request=GetLegendGraphic&TRANSPARENT=TRUE&LEGEND_OPTIONS=layout:verticle&format=image%2Fpng&width=20&height=20&layer=" +
                     layerName + 
-                    "&style=" + style;
+                    "&sld_body=" + currentStyle;
+                    //"&style=" + style;
     
         // all set - show the legend
         legendVisibilty = "visible";
