+++ src/utils/index.js
@@ -1,1 +1,2 @@
 export * from "./map-utils";
+export * from "./style-utils";
